tests/wpt/web-platform-tests/service-workers/service-worker/xhr.https.html

<!DOCTYPE html>
<title>Service Worker: XHR doesn't exist</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js?pipe=sub"></script>
<script>

async_test(function(t) {
    var path = new URL(".", window.location).pathname
    var url = 'resources/xhr.js';
    var scope = 'resources/blank.html?xhr';
    var host_info = get_host_info();
    var frameURL = host_info['HTTPS_ORIGIN'] + path + scope;

    service_worker_unregister_and_register(t, url, scope)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() { return with_iframe(frameURL); })
      .then(function(frame) {
        return new Promise(function(resolve, reject) {
          function onMessage(e) {
            assert_false(e.data.xhr);
            frame.remove();
            service_worker_unregister_and_done(t, scope);
          }
          var channel = new MessageChannel();
          channel.port1.onmessage = t.step_func(onMessage);
          frame.contentWindow.navigator.serviceWorker.controller.postMessage({port: channel.port2}, [channel.port2]);
        })
      })
      .catch(unreached_rejection(t));
  }, 'Verify XHR does not exist');
</script>
